:art: Move One functionality to API utility
