+++ lib/api/apis.js
@@ -188,26 +188,7 @@ const ofType = (query, type) => {
 const one = (query, id, model) => {
   // const organize = utils.organize(query);
 
-  return database('live').select('*').where({
-    id,
-  }).then(rows => {
-    const formatted = utils.format(rows, model)[0];
-
-    if (formatted) {
-      const type = {
-        name: formatted.type,
-        slug: formatted.type_slug,
-        url: `/api/types/${formatted.type_slug}`,
-      };
-
-      formatted.type = type;
-      delete formatted.type_slug;
-
-      return formatted;
-    }
-
-    return {};
-  });
+  return utils.one(query, id, model);
 };
 
 module.exports = {
