+++ lib/api/utils.js
@@ -2,6 +2,7 @@
 
 const _ = require('lodash');
 
+const database = require('../database');
 const schemas = require('../database/schemas');
 const utils = require('../utils');
 
@@ -234,9 +235,45 @@ const page = (base, organized, count) => {
   return pages;
 };
 
+/*
+ *  A Specific Piece of Content
+ *
+ * @param {object} query - Request queries
+ * @param {string} id - The GUID to search
+ * @param {array} model - Array of content type attributes
+ *
+ * @returns {object} - All items and the pagination for the items
+ */
+// TODO: Return error if no items are found
+const one = (query, id, model) => {
+  // const organize = utils.organize(query);
+
+  return database('live').select('*').where({
+    id,
+  }).then(rows => {
+    const formatted = format(rows, model)[0];
+
+    if (formatted) {
+      const type = {
+        name: formatted.type,
+        slug: formatted.type_slug,
+        url: `/api/types/${formatted.type_slug}`,
+      };
+
+      formatted.type = type;
+      delete formatted.type_slug;
+
+      return formatted;
+    }
+
+    return {};
+  });
+};
+
 module.exports = {
   attributes,
   format,
   organize,
   page,
+  one,
 };
